HPCC-23354 Supply an example helm chart and other docs
Changes following review

+++ dockerfiles/cleanup.sh
@@ -2,7 +2,7 @@
 
 # clean up old images
 
-docker rm $(docker ps -q -f 'status=exited')
+[[ "$1" == "-all" ]] && docker rm $(docker ps -q -f 'status=exited')
 docker rmi $(docker images -q -f "dangling=true")
 
 HEAD=$(git rev-parse --short HEAD)
